+++ README.md
@@ -7,13 +7,14 @@ Library for talking to the eos api.
 ```html
 <pre style="width: 100%; height: 100%; margin:0px; "></pre>
 
-<script src='.../eosjs2-debug.js'></script>
-<script src='.../eosjs2-jssig-debug.js'></script>
+<script src='dist/eosjs2-debug.js'></script>
+<script src='dist/eosjs2-jsonrpc-debug.js'></script>
+<script src='dist/eosjs2-jssig-debug.js'></script>
 <script>
     let pre = document.getElementsByTagName('pre')[0];
-    let endpoint = 'http://localhost:8000';
+    let rpc = new eosjs2_jsonrpc.JsonRpc('http://localhost:8000');
     let signatureProvider = new eosjs2_jssig.default(['5JtUScZK2XEp3g9gh7F8bwtPTRAkASmNrrftmx4AxDKD5K4zDnr']);
-    let api = new eosjs2.Api({ endpoint, signatureProvider });
+    let api = new eosjs2.Api({ rpc, signatureProvider });
 
     (async () => {
         try {
