+++ tsconfig.json
@@ -8,6 +8,7 @@
     },
     "files": [
         "src/index.ts",
+        "src/eosjs2-jsonrpc.ts",
         "src/eosjs2-jssig.ts"
     ]
 }
